requirements.txt
fastapi
uvicorn[standard]
jinja2
python-frontmatter
markdown
pygments